refactor(desktop): move App.tsx's module-scope components into components/
App.tsx 2855 -> 2374. Twelve components and the constant one of them owns move
into four files beside the four already in components/.

They are declared at module scope, so closing over App state was structurally
impossible and this is a move tsc fully checks. Three things were not
mechanical:

rowLabelClass and rowDescClass are read by SettingRow, SettingField AND App's
own Advanced disclosure, so leaving them behind would have meant
components/ importing from ../App: a cycle tsc resolves and module-init order
decides, with an undefined class string as the failure. They move to
components/ui.tsx, which already holds cn and the button variants, and are
exported.

dropVar travels with Dropzone, its only consumer, and stays unexported.

id="floe-undo-clear" was the real hazard. App.clearStaged() focuses it through
getElementById, so once UndoToast lived in another file a rename on either
side would compile clean, pass all 132 tests, and leave the keyboard falling
to document.body after a Clear. It is now an exported UNDO_ANCHOR_ID that both
sides import, which turns that drift into a compile error: renaming it fails
tsc in two places, verified by doing it. UndoToast had no test of any kind, so
components/Toasts.test.tsx now also pins what tsc still cannot see, that the
id lands on something focusable rather than a div.

Eleven imports went dead in App.tsx and nothing here would have said so: the
desktop frontend sets neither noUnusedLocals nor noUnusedParameters and has no
ESLint config. All eleven are pruned, and the two component/ui imports the
move created are merged back into one.

MutableRefObject is left alone. It is dead too, but it was already dead before
this PR, so removing it is a separate change.

frontend/wailsjs/go/main/App.d.ts is byte-identical, as it must be: no Go
changed and Wails reflects over the package-scoped method set.
